config: refactor into multiple files
