+++ lib/config/underwriter.go
@@ -1,0 +1,6 @@
+package config
+
+// UnderwriterOpts represents the underwriters info/models options.
+type UnderwriterOpts struct {
+	IncludeUnreleased bool `json:"include_unreleased"`
+}
